REFACTOR: gera_password no utils
